requirements/production.txt
dj-database-url==0.4.2
Django==1.10.5
djangorestframework==3.5.4
prettyconf==1.2.3
psycopg2==2.6.2
whitenoise==3.3.0
django-mptt==0.8.7
shortuuid==0.5.0